move everything to common folder
