+++ src/lib/precondition.ts
@@ -7,8 +7,8 @@ import {
   Preconditions,
 } from './account_update.js';
 import { UInt32, UInt64 } from './int.js';
-import { Layout } from '../snarky/gen/transaction.js';
-import { jsLayout } from '../snarky/gen/js-layout.js';
+import { Layout } from '../provable/gen/transaction.js';
+import { jsLayout } from '../provable/gen/js-layout.js';
 import { emptyReceiptChainHash } from './hash.js';
 import { PublicKey } from './signature.js';
